Update GetAccountTransactions implementation according to spec

+++ client/mock_client.go
@@ -143,14 +143,14 @@ func (_m *MockAptosClient) GetAccountResources(ctx context.Context, address stri
 }
 
 // GetAccountTransactions provides a mock function with given fields: ctx, address, start, limit, opts
-func (_m *MockAptosClient) GetAccountTransactions(ctx context.Context, address string, start int, limit int, opts ...interface{}) ([]TransactionResp, error) {
+func (_m *MockAptosClient) GetAccountTransactions(ctx context.Context, address string, start uint64, limit int, opts ...interface{}) ([]TransactionResp, error) {
 	var _ca []interface{}
 	_ca = append(_ca, ctx, address, start, limit)
 	_ca = append(_ca, opts...)
 	ret := _m.Called(_ca...)
 
 	var r0 []TransactionResp
-	if rf, ok := ret.Get(0).(func(context.Context, string, int, int, ...interface{}) []TransactionResp); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, string, uint64, int, ...interface{}) []TransactionResp); ok {
 		r0 = rf(ctx, address, start, limit, opts...)
 	} else {
 		if ret.Get(0) != nil {
@@ -159,7 +159,7 @@ func (_m *MockAptosClient) GetAccountTransactions(ctx context.Context, address s
 	}
 
 	var r1 error
-	if rf, ok := ret.Get(1).(func(context.Context, string, int, int, ...interface{}) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, string, uint64, int, ...interface{}) error); ok {
 		r1 = rf(ctx, address, start, limit, opts...)
 	} else {
 		r1 = ret.Error(1)
